semantic_search/requirements.txt
mysql-connector-python
sentence-transformers
pandas
faiss-cpu 
numpy